rag_system/requirements.txt
openai==1.57.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
langgraph==0.2.59
langchain-openai==0.2.14
langchain-core==0.3.28
supabase==2.10.0
python-dateutil==2.8.2
